refactor(service-layer): #34 extracted rules related handling method a separate class
